requirements.txt
streamlit==1.31.0
# Using specific version matching for LangChain packages to ensure compatibility
langchain==0.1.12
langchain_openai==0.0.6
pydantic>=2.5.2
openai>=1.12.0
numpy==1.26.3
pandas==2.2.0
PyPDF2==3.0.1
# Using specific version for LangGraph to ensure compatibility
langgraph==0.0.16
# Vector database for document similarity search
chromadb>=0.4.24
matplotlib
# Voice functionality
openai-whisper
ffmpeg-python
pyaudio
audio_recorder_streamlit
pypdf2
python-dotenv
